Add syntax for volatile fields Move all "small"/"helper" methods from ApiGenerator to extensions to keep it clean fixes #116
